3.2.23 (2015-1-30) - Corrects the sideband reversal when using the SAM sideband select feature. - Corrects the NaN error in the SWR display. - Corrects error in reporting negative 6m reverse power values. - Added workaround for initializing Hermes PA relays. - Added "Stereo Diversity" capability for Beta testing.

+++ trunk/W5WC/PowerSDR_HPSDR_mRX_PS/Source/Console/DJConsoleUI/DJConsole.cs
@@ -3632,7 +3632,7 @@ namespace PowerSDR
             parser.nGet = 0;
             int step = StringToFreq(commands.ZZAC(""));
             parser.nSet = 11;
-            long freq = Convert.ToInt32(commands.ZZFA(""));
+            long freq = Convert.ToInt64(commands.ZZFA(""));
             parser.nAns = 11;
 
             int mode = Convert.ToInt16(commands.ZZMD(""));
@@ -3758,7 +3758,7 @@ namespace PowerSDR
             parser.nGet = 0;
             int step = StringToFreq(commands.ZZAC(""));
             parser.nSet = 11;
-            long freq = Convert.ToInt32(commands.ZZFB(""));
+            long freq = Convert.ToInt64(commands.ZZFB(""));
             parser.nAns = 11;
 
             int mode = Convert.ToInt16(commands.ZZMD(""));
